Refactor variable and loading system
Add a dist/release.sh script that generates a zip file of the release
that can be uploaded directly to DWC.

Allow MOS to be reloaded on-the-fly. NOTE: Non-user-vars are not
reloaded as they can only be defined once, so if MOS variables are added
or updated in that file then RRF needs reloading.

Note: Macros are not in a runnable state in this commit, they need work
to follow along with the format.

Signed-off-by: Ben Agricola <ben+git@agrico.la>