Add meta macro for "probe work piece" and post
This is starting to build up a (non-working, as of right now) structure
of how we will use the post-processor to control workpiece probing.

The post will output a G6600 command when it wants a workpiece to be
probed. This will ask the operator to select a probing operation (e.g.
Bore, Pocket, Outside Corner), and will execute a further G65xx command
to ask the operator for the parameters required to execute the probing
op.

The corresponding probing op will then be executed (e.g. G65xx.1), and
the work co-ordinate offset will be set.

The G6600 command will return once the probing operation is complete,
allowing the program to continue.

Signed-off-by: Ben Agricola <ben+git@agrico.la>