Add protected moves
Still working using Circular Bore as the test macro, we now implement
G6550.1 and G6550.2 which are protected movement operations.

G6550.1 moves towards a target location, observing for probe activation.
If the probe is activated it stops, and we then compare the target position
with the current position to work out if we reached the target position
or were stopped early due to a collision. We compare positions using a
0.1mm accuracy since this is generally accurate enough for probe start
positions.

G6550.2 moves towards a target location, observing for probe
deactivation.
When the probe is deactivated it stops. We can then switch to using a
G6550.1 to move the rest of the distance. This is useful when backing
off from a workpiece we have just probed - G6550.1 would abort instantly
as the probe is activated at the start of the move.

Signed-off-by: Ben Agricola <ben+git@agrico.la>