Structural changes and Config Wizard
Okay, this is a pretty big commit (whoops!) The main takeaways from this
commit should be a bunch of new features aimed a new machinists.

First off, the first time you start MillenniumOS you will now be walked
through a configuration wizard. This wizard works out where your
configured touch probe and toolsetter are and which spindle you're
using. It then takes a bunch of measurements that are required for
specific functionality during machining, like waiting for the spindle to
accelerate and decelerate, and making accurate tool changes and work
piece probing.

You can run the config wizard macro manually using G8000 or clicking it
in the Macro list.

Secondly, a number of public macros have been moved into a folderised
structure to make their usage more clear.

This now includes a bunch of new higher-level probing macros that are in
various states of disarray. Do _not_ use these yet unless you're happy
working out why they launched your machine into next tuesday. Highlights
are macros for spindle control that wait for the spindle to reach the
requested speed (M3.1 and M5.1 - TBD, we might move all macros on
existing codes to .9 or higher to avoid potential conflicts if RRF adds
new decimal codes).

We also add M25.1 which pauses processing of a gcode file if a gcode
file is being executed, otherwise it simply aborts. This is used during
tool changes to pause the job if an error has occurred rather than
aborting it entirely, allowing the operator to recover the issue.

Finally we add M6, which is used to execute a tool change based on the
configured tool information.

Signed-off-by: Ben Agricola <ben+git@agrico.la>