Add tool radius compensation to probe points
Noticed during recording a video that when using a high tool radius
(e.g. a 6mm dowel) that the probe start and target points do not take
this into account - with a 10mm clearance, the surface of a 6mm dowel
should be about 7mm from the surface to be probed - but the manual probe
distance to target was 12mm. This would allow the operator to move the
_center_ of the 6mm dowel 2mm inside the surface, which is obviously way
too much.

We now modify the clearance and overtravel values to account for this,
increasing the clearance by the tool radius and decreasing the
overtravel by the same. In some cases, this means the overtravel becomes
zero or a negative value but this is fine, because when taking into
account the outer surface of the probe tool, it would still be against
or inside the surface we're trying to probe.

Other changes:

 - Provide a little more feedback on "Unknown Error" at startup.
 - Add a section to the README to describe how to configure Z-probes in
   RRF.
 - Add a "Back-Off" option to manual probes, so that if the operator
   accidentally jogs too far they can back off by 1mm and approach
   again.

Signed-off-by: Ben Agricola <ben+git@agrico.la>