Apply probe tool radius to probe points
We did not previously account for the tool radius in our probe start and
target points. We now apply the tool radius to the clearance and
overtravel in each higher-level probing macro to make sure that when we
specify a clearance of say, 10mm, then the _surface_ of the probing tool
should end up approximately 10mm from the target location - rather than
this being 10mm from the _center_ of the tool.

We now also make the W parameter when calling probing macros directly
optional.

The M7600 command to output known variables had not been updated after
re-working the probe macros to use the same variables for their output.

We also inline a number of checks around Tool selection (to reduce
memory usage).

Signed-off-by: Ben Agricola <ben+git@agrico.la>