Limit backoff distance during probes
It is possible for the operator-entered backoff distance to be higher
than the distance we actually moved from the starting point. In these
cases we want to limit the back-off distance so we can only ever move
back to the starting point. This should allow us to probe small bores
more safely.

Signed-off-by: Ben Agricola <ben+git@agrico.la>