Fix single surface probe distance
Single surface probes did not account for overtravel and probe radius
correctly. Since we can probe in both X/Y and Z we need to apply
overtravel for everything but radius only on X/Y probes, before we were
applying it to overtravel but not applying overtravel to any moves. This
is now fixed.

We also make sure that calls to G6512 include all 3 starting locations
so if a toolchange is triggered at the start of a probing procedure then
any position the user jogged to will be returned to after the
toolchange. This stops us probing in an unknown location.

We move touch probe detection response to tpost.g. It is not
currently possible to abort a toolchange procedure (this needs changing
in RRF if possible) so currently the tool number does not act 100%
safely as a guard for touchprobe operations, but if running as part of a
file then the file will be successfully aborted if the touch probe is
not detected.

We fix some input box issues where a probed radius was inserted into a
diameter input field, instead of multiplying it by 2.

We remove partly-implemented probing routines and clean up some variable
assignments.

We also update the readme with some warnings and a line about
post-processor installation.

Signed-off-by: Ben Agricola <ben+git@agrico.la>