Skip null probe issues in M8001
If undefined, probe sensors will be null within the sensors.probes
vector. Specifying toolsetter as probe ID 1 and touch probe as 2 means
there will be a single 'null' entry at 0 that we were trying to look for
a type variable on. We skip all null probes now so M8001 should work as
expected if probe 0 is not defined.

We also revert a change to global.mosToolTable that turned it from a
vector with one item into a scalar. TL;DR variables in RRF must be
appended with a comma to create a single item vector.

Signed-off-by: Ben Agricola <ben+git@agrico.la>