Fix wizard and invalid tool index issues
RC1 did as expected - identified a bunch of issues on first-contact with
a real user. Unfortunately they were in the Wizard process, the first
interaction that a new user would have with MOS 🤦

We fix these issues in various different manners, but mostly by checking
to make sure that a tool is currently selected before trying to read the
tool radius. If no tool is selected then we do not perform radius
compensation during probing operations.

To get around tool limitations in the Wizard (before MOS is properly
configured, which is necessary for correct tool changes), we temporarily
create and delete probe tools so we can run the steps necessary to
calibrate the touch probe, datum tool and toolsetter.

Finally, we add some checks so that Toolsetter and Touch Probe cannot be
toggled if they have not yet been configured in the Wizard.

Many thanks to @MrSniperPhil for being the guineapig and putting his
Milo and probe tips on the line for this!

Signed-off-by: Ben Agricola <ben+git@agrico.la>