Validate RRF tool table indexing
It turns out the RRF tool table is lazily allocated. If you define a
tool with index 1, then RRF will create a table with 2 entries (0 and 1,
with 0 being null). This means it is not simply enough to check that the
tool number is within the limits, because the tool entry itself might
not exist.

This did not show up in testing because we use the last index in the
tool table as the probe tool, and our RRF config was not creating tools
in config.g.

It is still good practice to not create tools in config.g as MOS is
supposed to manage the tool list, but we should no longer throw errors
when trying to add or remove tools using M4000 and M4001 if the tool
table has been lazily allocated.

Also update M7600 so it outputs useful object model fields for debugging
when called with D1.

Signed-off-by: Ben Agricola <ben+git@agrico.la>