Fix usage of exists() on vectors
Usage of `exists(vector[n])` in RRF is apparently undefined behaviour,
and it turns out this actually returns the value of the index of the
vector rather than checking if it exists. This is very confusing when
the vector contains booleans!

We replace all usage of exists() on vectors by checking the index into
the vector is less than the length of the vector, and greater than zero.

Signed-off-by: Ben Agricola <ben+git@agrico.la>