Compensate for probe deflection in X and Y separately
Previously, we were averaging the deflection value between X and Y and
applying the same deflection compensation to both axes.

This commit splits the deflection value into a vector, first in the
configuration wizard, and then in the tool management code (M4000).

Finally, the tool-stored deflection value is applied to the probe
location in both X and Y, multiplied by the magnitude of the direction
vector of the probe movement - i.e. adjusted based on the direction of
the probe move.

Signed-off-by: Ben Agricola <ben+git@agrico.la>