Allow wizard reconfig of feature settings
If MillenniumOS is already loaded, the configuration wizard will now
prompt to ask if you would like to reconfigure the spindle, datum tool,
toolsetter and touch probe. If you click yes to any of these, the
relevant settings will be nullified and you will be able to configure
all of them again.

We also fix a bug in the release script that did not appear to be
replacing the %%MOS_VERSION%% placeholder correctly.

Finally, we change the Fusion360 post-processor so that work offsets in
Fusion360 match the default Duet Web Control values.

Signed-off-by: Ben Agricola <ben+git@agrico.la>