Add G6502 Rectangle Pocket probe routine
This commit adds the last remaining probing routing required before a
beta release - Rectangle Pocket. It is based on the G6503 Rectangle
Block routine but probes outwards instead of inwards.

There are also some small supporting changes in this commit and a
warning in the Reload macro that users must _restart_ their mainboard
after uploading a new MillenniumOS release, as a reload might not
account for all changes (default variable values in mos-vars.g are not
reloaded currently as they might override manually reconfigured values).

Signed-off-by: Ben Agricola <ben+git@agrico.la>