Fix protected move back-off distances
Previously, the back-off distance (when probe is activated and we need
to back off slightly so we can move the rest of the way looking for
probe activation) was based on the tool radius, but this is open to
issues with very large or very small tools. Instead, we add a new
setting, `global.mosProtectedMoveBackoff` which is a distance in mm to
back off before trying a protected move. This setting is configurable in
the configuration wizard.

We also fix reference surface probes.

Reference surface probes are needed if a touch probe has been used to
set workpiece origins. Even if we do not output probe cycle commands
from the post (because the origin is already probed), we _must_ make
sure that the surface position has been probed otherwise tool changes
will generate an error.

We now make sure `G6511` can be called from a gcode file. It will be a
no-op if the surface is already probed or the toolsetter or touch probe
feature is disabled. It will switch to the touch probe if necessary.

It is run before any generated probe cycles so any initial WCS probe
will not require disconnecting and reinserting the touch probe.

We make sure `G6511` is outputted by the post-processor even in WCS
probe mode NONE.

We fix an issue with `M4000` that would fail if RRF tools existed but
were null. We also now allow `M4000` to overwrite existing tools, they
no longer need to be reset using `M4001` before calling `M4000`.

We also update the README to better explain how to install MillenniumOS
using the Duet Web Console.

Finally, there was an issue with VSSC and an incorrectly named variable
which is now fixed.

Signed-off-by: Ben Agricola <ben+git@agrico.la>