Fix RC9 issues
Phil found a bunch of issues with RC9 which involved the long-standing
floating-point issue which should now be fixed by checking for results
in a range rather than ceil / flooring and looking for equality.

We also found an issue with VSSC where it sets the spindle speed for
tool 0 which is now incorrect given that we use multiple tools. We
remove the tool ID so it now sets the speed of the current tool.

We move mosProtectedMoveBackOff above touch probe and toolsetter
movements in the config wizard as it is required to be set by these
systems.

Finally, we add an M598 call and make sure to exit macros if not running
in the primary movement system, which should help with some odd macro
behaviour.

Signed-off-by: Ben Agricola <ben+git@agrico.la>