RC9 Fixes (#20)
* Fix RC9 issues

Phil found a bunch of issues with RC9 which involved the long-standing
floating-point issue which should now be fixed by checking for results
in a range rather than ceil / flooring and looking for equality.

We also found an issue with VSSC where it sets the spindle speed for
tool 0 which is now incorrect given that we use multiple tools. We
remove the tool ID so it now sets the speed of the current tool.

We move mosProtectedMoveBackOff above touch probe and toolsetter
movements in the config wizard as it is required to be set by these
systems.

Finally, we add an M598 call and make sure to exit macros if not running
in the primary movement system, which should help with some odd macro
behaviour.

* Comment all debug lines

This reduces memory usage significantly.

Also switch the protected move back-off to not be protected.

This is a choice to avoid back-off issues when the probe reacts too fast
and bounces.

We also define an undefined variables for VSSC.

* Remove nesting for G6520 to test

This removes the M291 nesting that was necessary with multiple motion
systems active.

This also changes some behaviour during tool changes, to avoid using
interpolated strings and making sure they're not run in multiple motion
systems.