Reduce heap memory usage (#21)
This is a combination of changes that hits just about everything -
sorry... again.

We shorten variable names so that they take up as little heap space as
possible while still being understandable.

Some variables are a little more verbose, these are the ones a user
might interact with, like the WorkPiece (WP) dimensions, angle,
co-ordinates etc.

We reduce the use of vectors that contain non-string, non-array values
since these are always stored on the heap. We now use multiple variables
for some of these, usually post-fixed with X, Y and Z for co-ordinates
or distances.

We implement a tracking value for tool changes, so that we do not
blindly try to run a tool probe when there was an error in the
pre-tool-change macro. RRF executes 'system' macros unconditionally so
we must make sure we abort the current job if an error occurs during the
tool change, and make sure subsequent tool change macros exit early if
so.

We also reduce the number of global variables in some situations (where
they are only used 'locally') and we make sure to use variables from the
object model if we have them available, rather than calculating from
other sources.

Signed-off-by: Ben Agricola <ben+git@agrico.la>