Make sure expert mode does not put fingers at risk
There were a couple of locations where turning on Expert mode (`global.EM
= true`) would trigger a move straight after the user had been asked to
install or activate the touch probe or datum tool with no confirmation.
Given that operator fingers would be within the vicinity of the machine,
this infers a (small - because the table moves not the spindle - but
real) risk of injury, and so we have now removed the skipping of those
confirmations within the expert mode feature.

There are still a small number of places where expert mode skips user
confirmations but these are after the operator should have already moved
away from the machine.

We also fix a small number of spelling errors and add some additional
information.

Signed-off-by: Ben Agricola <ben+git@agrico.la>