Add standard and FMJ F360 machine definitions
Annoyingly, Fusion360 won't show operation properties unless
the post-processor has been assigned to a machine. So here we add 2
machine definitions, one for the standard dimensions, and one for the
extended Z used by the kits with FMJ and 120mm clearance.

We also slightly modify the output of the post-processor to make it
easier to read, and add comments on more lines. Finally, we clean up
some markdownlint issues in the post-processor readme.

Also add the workflow steps to upload our machine files to new releases.

Signed-off-by: Ben Agricola <ben+git@agrico.la>