Fix machine limit error if moving too close to ref surface during G8000
Fixes #28 by using the lower of abs(minimum z pos - current z pos) or
20mm as the probing distance, so this should always be within the Z
limits of the machine

We also remove the overtravel value as this would still allow us to
trigger an out of limits message in certain situations.

This way, we should always be able to move to _at least_ the axis
minimum if the machine position is within 20mm of it.

We also remove some unnecessary variables in `G6510.1`, falling back to
using the `param.J`, `param.K` and `param.L` values directly as they are
only used once and not calculated.

Signed-off-by: Ben Agricola <ben+git@agrico.la>