Fix resume move
We should move to the X/Y position above the resume state before moving
only down in Z, this avoids diagonal moves.

Additionally, we should deselect the active tool at the end of a job,
and allow `tpre.g` to run even if the previous `tpre.g` run failed. This
allows us to bring the machine back into a consistent state. `tpre.g`
should only abort if `tfree.g` was run and errored.

Finally, we remove the ability to pause after a tool change - this
functionality is broken in RRF and leaves the machine in an unresumable
state, requiring a restart.

Signed-off-by: Ben Agricola <ben+git@agrico.la>