Fix target outside machine limit error in resume.g
Since we tried to move a relative distance 10 above the stored resume
position, it's possible that pausing at Z > -10 would try to move the
machine above Z max and trigger an error.

This error would not abort the job or cause any damage but is
distracting.

Since we park the spindle in Z at the start of the resume process, we
know the spindle is already at Z max - we can move it over the resume
position and then plunge directly to the stored location, avoiding the
potential error message.

Signed-off-by: Ben Agricola <ben+git@agrico.la>