Make sure Spindle start can be paused
Previously, clicking cancel on the "Spindle Start" confirmation dialog
would not start the spindle but _would_ continue with the running job,
potentially breaking your tool. Rather than use the inbuilt "OK" /
"Cancel" dialog, we not implement a "Continue" / "Pause" dialog that
appears if the machine is not currently pausing, paused or resuming when
`M3.9` is called.

Remember: Expert mode bypasses this check so you should be well away
from the machine when using expert mode and completing a tool change, as
it is very likely that `M3.9` is run directly after.

Signed-off-by: Ben Agricola <ben+git@agrico.la>