Fix resume moves and tool change issues (#30)
* Fix resume move

We should move to the X/Y position above the resume state before moving
only down in Z, this avoids diagonal moves.

Additionally, we should deselect the active tool at the end of a job,
and allow `tpre.g` to run even if the previous `tpre.g` run failed. This
allows us to bring the machine back into a consistent state. `tpre.g`
should only abort if `tfree.g` was run and errored.

Finally, we remove the ability to pause after a tool change - this
functionality is broken in RRF and leaves the machine in an unresumable
state, requiring a restart.

Signed-off-by: Ben Agricola <ben+git@agrico.la>

* Minor tool-change fixes

Simple changes to error or warning messages encountered during a tool
change to make them clearer. We also re-order some of the 'abort'
conditionals to make sure we exit as early as possible if the machine
is not in an appropriate state to change tools.

Signed-off-by: Ben Agricola <ben+git@agrico.la>

* Fix tpre after tfree failure abort message

Signed-off-by: Ben Agricola <ben+git@agrico.la>

* Check for null tool change state in tpost

If `tpost.g` were called without `tfree.g` and `tpre.g` executing far
enough to set the tool change state, then we could throw an error
attempting to compare null to < 3. We now explicitly check for null and
abort the `tpost.g` file.

Signed-off-by: Ben Agricola <ben+git@agrico.la>

* Fix target outside machine limit error in resume.g

Since we tried to move a relative distance 10 above the stored resume
position, it's possible that pausing at Z > -10 would try to move the
machine above Z max and trigger an error.

This error would not abort the job or cause any damage but is
distracting.

Since we park the spindle in Z at the start of the resume process, we
know the spindle is already at Z max - we can move it over the resume
position and then plunge directly to the stored location, avoiding the
potential error message.

Signed-off-by: Ben Agricola <ben+git@agrico.la>

* Make sure Spindle start can be paused

Previously, clicking cancel on the "Spindle Start" confirmation dialog
would not start the spindle but _would_ continue with the running job,
potentially breaking your tool. Rather than use the inbuilt "OK" /
"Cancel" dialog, we not implement a "Continue" / "Pause" dialog that
appears if the machine is not currently pausing, paused or resuming when
`M3.9` is called.

Remember: Expert mode bypasses this check so you should be well away
from the machine when using expert mode and completing a tool change, as
it is very likely that `M3.9` is run directly after.

Signed-off-by: Ben Agricola <ben+git@agrico.la>

---------

Signed-off-by: Ben Agricola <ben+git@agrico.la>