Reimplement VSSC using a sine function (#31)
The original implementation of VSSC was a bit naiive. Using a sine
function to generate a number between 0 and 1, and then adjusting the
speed from the lower limit forms a smooth adjustment that can be applied
at every run of `daemon.g` instead of running a single RPM change to the
top and the bottom of the limit.

This helps where VFD's are set with very fast acceleration and
deceleration values, to avoid very abrupt changes in speed.

However, this does change the implementation slightly. The VSSC period
is now the time duration of a single wavelength, and the VSSC variance
is now twice the amplitude of the wave. So calling M7000 P2000 V500
would vary the spindle speed from the lower limit to the upper limit and
back in 2000ms, where the lower limit is 250rpm below the base RPM and
the upper limit is 250rpm above it.

This is unlikely to have significant implications for existing usage,
but if a reduction in surface quality or increase in chatter is noticed,
you should tweak your VSSC settings. The F360 post-processor has been
updated by doubling the VSSC period and variance so it still applies
roughly the same spindle speed adjustments as it defaulted to before.

Signed-off-by: Ben Agricola <ben+git@agrico.la>