Refactor manual probe macro and add low-mem mode
Various changes attempting to reduce memory usage during manual probing.
We try to allocate vectors only once and reduce the number of variables
in use.

We also add a low memory mode to the Fusion360 post-processor which
linearizes all arc moves.

There appears to be a heap allocation required to process arc moves
within RRF (likely so the move can be segmented into linear moves
internally or similar) and if we're running with very low free heap
memory, this will cause a heap allocation and an OutOfMemory error.

Low Memory mode should avoid this allocation by not using G2 or G3
commands.

Signed-off-by: Ben Agricola <ben+git@agrico.la>