Fix divide by zeros
There's a number of places where we calculate direction vectors. It's
possible for these to return zero if there is no movement along either
axis (start equals target), or if a surface is perfectly parallel to an
axis. In these cases, we trigger a divide by zero and this has very
unintended consequences.

Signed-off-by: Ben Agricola <ben+git@agrico.la>