Fix nan detection in outside corner probe
Previously, we were checking if mX and mY vars were zero, but what we
actually wanted was to check if they were nan.

Also do not apply X/Y tool radius compensation if the probe was not
moved in those axes (for example, when using the probing code to set a
tool length). This avoids triggering nan detection where we didn't even
attempt to move the probe.

Finally, make sure most M-codes do not run in the secondary motion
system so we do not end up running various things twice (like spindle
control).

This comes with the relevant documentation updates, as well as making it
clear that operators should follow the documentation on our dedicated
site and that the README contents are mostly for advanced users.

Signed-off-by: Ben Agricola <ben+git@agrico.la>