Rework manual probing, fix various divide-by-zero / nan issues, add a low-memory mode for the F360 post-processor and support drilling cycles. (#34)
* Refactor manual probe macro and add low-mem mode

Various changes attempting to reduce memory usage during manual probing.
We try to allocate vectors only once and reduce the number of variables
in use.

We also add a low memory mode to the Fusion360 post-processor which
linearizes all arc moves.

There appears to be a heap allocation required to process arc moves
within RRF (likely so the move can be segmented into linear moves
internally or similar) and if we're running with very low free heap
memory, this will cause a heap allocation and an OutOfMemory error.

Low Memory mode should avoid this allocation by not using G2 or G3
commands.

* Add units to end of VSSC setting field names

* Fix divide by zeros

There's a number of places where we calculate direction vectors. It's
possible for these to return zero if there is no movement along either
axis (start equals target), or if a surface is perfectly parallel to an
axis. In these cases, we trigger a divide by zero and this has very
unintended consequences.

* Add divide by zero / nan checks

There are various places that calculate a slope or normal between two
points. If any of these calculations return 0, then subsequent
calculations where this input value is divided would return a nan.
Setting a WCS to nan causes odd stuff in RRF and also breaks
calculations.

* Add drilling cycle support to F360 post

A number of drilling cycles are now supported in the Fusion360
post-processor, alongside the ability to display an operator message
using the "Manual NC" command set to "Display Message".

* Fix nan detection in outside corner probe

Previously, we were checking if mX and mY vars were zero, but what we
actually wanted was to check if they were nan.

Also do not apply X/Y tool radius compensation if the probe was not
moved in those axes (for example, when using the probing code to set a
tool length). This avoids triggering nan detection where we didn't even
attempt to move the probe.

Finally, make sure most M-codes do not run in the secondary motion
system so we do not end up running various things twice (like spindle
control).

This comes with the relevant documentation updates, as well as making it
clear that operators should follow the documentation on our dedicated
site and that the README contents are mostly for advanced users.