Add Z reset procedure for toolsetter-less changes (#40)
We add G37.1 which is a Z surface probing procedure that does not
change to the probing tool before running a manual-only probing
procedure.

We integrate this into the tool change macros so if the toolsetter is
not enabled, the Z origin can be reset at each tool change. MillenniumOS
can now be used without a touch probe or a toolsetter!

We also try to reduce the usage of the S3 dialog mode as clicking Cancel
on this dialog does not pass the cancel back to the calling macro, and
might continue processing where we would expect it to... not. We still
have to use it when expecting the operator to jog, because jogging is
only supported on S<=3 modes, but we are now 'safer' where we can be.

We now only probe the reference surface if both the toolsetter and touch
probe feature are enabled.

We fix a number of tutorial mode prompts and make some copy changes to
make these clearer.

Signed-off-by: Ben Agricola <ben+git@agrico.la>