Fix `M3.9` and `M5.9` operator warnings and dwells
We now validate all possible input parameters to `M3.9` and `M5.9`,
calculate the dwell time based on the percentage of the maximum spindle
speed of the change, and only warn the operator during `M3.9` if the
spindle is accelerating from a stop.

Additionally, with `M5.9` we will adjust the delay based on the speed of
the spindle that was deemed to be active, if there was one active.

Signed-off-by: Ben Agricola <ben+git@agrico.la>