Fix error messages from VSSC at max RPM
We did not calculate the lower (and therefore upper) limit for VSSC
correctly when the target RPM was less than the variance from the
spindle maximum. We would throw errors while machining and not adjust
the speed upwards. Although this would not stop the job running, it does
fill the logs with spam.

Signed-off-by: Ben Agricola <ben+git@agrico.la>