Fix probes not throwing errors when deactivated
Probe cycles would continue attempting to probe if a probe point was not
activated - this was down to the 'errors' functionality added to support
radius offset tool probing.

When the `E` param was not specified, `null` would be passed to the
underlying probing macro - but we were using `if(!exists(param.E))` to
validate this.

Passing `Enull` to RRF means that `param.E` _exists_, so that check does
not correctly turn on errors.

We fix this by making the following check `param.E != 0`, so we only
disable errors if the parameter is specifically 0. Then we _also_ pass
`1` in all cases where `param.E` was not given by the higher-level
probing macro.

Signed-off-by: Ben Agricola <ben+git@agrico.la>