Fix VSSC errors when no tool selected
When controlling the spindle manually via DWC with VSSC activated,
errors would be thrown every cycle of the daemon if no tool was
selected.

We now do not attempt to set the spindle RPM if no tool is selected.
This is necessary because of how we control the spindle speed based on
the tool speed.

Signed-off-by: Ben Agricola <ben+git@agrico.la>