Fix intermittent HardFault Unaligned on null
We were previously defaulting the deflection value to `null` in the tool
table - but this appears to take up half as much space as the actual
value, which ends up being 2 floats. In certain situations this would
cause a HardFault Unaligned and bootloop the mainboard.

By not using `null` as the default value, we should avoid overflowing
the tool table and boot as normal.

Signed-off-by: Ben Agricola <ben+git@agrico.la>