Implement storage of multiple WCS probing details
Rather than storing the center or corner of a probed workpiece in a
single variable, we now store these in a set of vectors based on their
WCS number. This way, this data can be referred to later if there are
multiple WCS in use.

This is important for implementing rotation compensation and likely
other things, but also has a significant negative impact on memory usage
which makes it very unlikely to work on current STM32F4 chips until
their memory allocation behaviour is changed.

Signed-off-by: Ben Agricola <ben+git@agrico.la>