Make sure to clear existing rotation compensation
Rotation compensation is not a workplace-specific setting in RRF, but we
treat it as such in MillenniumOS. If no rotation compensation value is
available, or the user specifically asks for the rotation compensation
to not be applied, then we should call `G69` to cancel any existing
rotation.

This makes sure that on switching to a new WCS, any existing rotation
applied on the previous WCS will be deactivated.

Signed-off-by: Ben Agricola <ben+git@agrico.la>