Set the toolsetter Z when probing datum tool
It seems there have been a large number of issues related to inaccurate
Z height or machines going out of limits. This appears to be down to
issues with how the Z offset was calculated, using the original position
of the toolsetter as probed during the wizard to calculate an offset for
the datum tool itself, and then for any subsequent tools.

We now set the toolsetter Z position when the datum tool is probed
rather than giving it an offset, and then set offsets on subsequent
tools based on the probed activation point.

Signed-off-by: Ben Agricola <ben+git@agrico.la>