Add mos-override-vars.g file
Allow MillenniumOS settings to be overridden after the MillenniumOS
user-vars file has been loaded. This can be used to set surface names
and corner names taking into account that some machines may not move in
the same directions as Milo by default.

This file will not be overwritten on upgrade.

Signed-off-by: Ben Agricola <ben+git@agrico.la>